tuhi: move the gui/svg.py file to tuhi
One JSON to SVG converter per project is enough.
